+++ packages/tools/pluggable-widgets-tools/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+- Updated the typings-generator with the proper types for widget properties of type `icon` and `image`.
+
 ## 9.2.1 - 2021-06-16
 - We reintroduced the possibility to use `.env` file for environment variables (removed in v9 previously)